requirements.txt
google-search-results==2.4.2
PyYAML==6.0.2
resend==2.10.0
python-dotenv==1.0.1
pytest==8.3.5
